requirements.txt
Flask==2.0.3
gensim==4.2.0
gevent==21.12.0
jieba==0.42.1
joblib==1.0.1
lightgbm==3.3.2
numpy==1.21.1
pandas==1.3.1
PyMySQL==1.0.2
PyYAML==6.0
scikit_learn==1.1.2
tqdm==4.62.0
